# Conflicts:
#	backend/src/main/java/com/coyjiv/isocial/transfer/friend/FriendNotificationMapper.java
#	backend/src/main/java/com/coyjiv/isocial/transfer/post/RepostNotificationMapper.java

+++ backend/src/main/java/com/coyjiv/isocial/service/like/ILikeService.java
@@ -14,7 +14,7 @@ public interface ILikeService {
 
   void toggleLike(Long userId, Long entityId, LikeableEntity entityType) throws EntityNotFoundException;
 
-  int countLikesByEntity(Long entityId, LikeableEntity entityType) throws EntityNotFoundException;
+  long countLikesByEntity(Long entityId, LikeableEntity entityType) throws EntityNotFoundException;
 
   boolean isLikedByUser(Long userId, Long entityId, LikeableEntity entityType) throws EntityNotFoundException;
